Fix augmentations?
