+++ src/data/generator.py
@@ -18,11 +18,15 @@ class DataGenerator(tf.keras.utils.Sequence):
                  batch_size,
                  height=64,
                  channels=1,
+                 augment_model=None,
+                 is_training=True,
                  ):
         self.batch_size = batch_size
         self.tokenizer = tokenizer
         self.height = height
         self.channels = channels
+        self.augment_model = augment_model
+        self.is_training = is_training
 
     def load_images(self, image_info_tuple: Tuple[str, str]) -> (
             Tuple)[np.ndarray, np.ndarray]:
@@ -68,8 +72,21 @@ class DataGenerator(tf.keras.utils.Sequence):
         image = tf.image.resize(image, (self.height, 99999),
                                 preserve_aspect_ratio=True) / 255.0
 
-        image_width = tf.shape(image)[1]
+        # Add batch dimension
+        image = tf.expand_dims(image, 0)
 
+        # Apply augmentations
+        if self.augment_model is not None:
+            for layer in self.augment_model.layers:
+                if not self.is_training and \
+                        isinstance(layer, tf.keras.layers.Cropping2D):
+                    continue
+                image = layer(image, training=self.is_training)
+
+        # Remove batch dimension
+        image = image[0]
+
+        image_width = tf.shape(image)[1]
         image = tf.image.resize_with_pad(image, self.height, image_width+50)
 
         label = image_info_tuple[1]
